refactor: flatten to root dir
