+++ src/fetcher.ts
@@ -1,7 +1,7 @@
-import { RequestMethod, MiddlewareResult } from '../interfaces';
-import { FatcherError } from '../FatcherError';
+import { RequestMethod, MiddlewareResult } from './interfaces';
+import { FatcherError } from './FatcherError';
 import { isPlainObject, Querystring } from '@fatcherjs/utils-shared';
-import { defineMiddleware } from '../helpers';
+import { defineMiddleware } from './helpers';
 
 /**
  * A middleware for send http request by using fetch.
