--- src/middlewares/index.ts
@@ -1,3 +1,0 @@
-export * from './composeMiddlewares';
-export * from './fetcher';
-export * from './registerMiddlewares';
